refactor(cel): Move cel conversion functions to separate package
Moves conversion functions into pkg/cel/conversion package to allow cel libraries to import without circular dependencies
